+++ Forge/src/main/java/at/petrak/hexcasting/forge/cap/adimpl/CapItemMediaHolder.java
@@ -11,17 +11,17 @@ public record CapItemMediaHolder(MediaHolderItem holder,
                                  ItemStack stack) implements ADMediaHolder {
 
     @Override
-    public int getMedia() {
+    public long getMedia() {
         return holder.getMedia(stack);
     }
 
     @Override
-    public int getMaxMedia() {
+    public long getMaxMedia() {
         return holder.getMaxMedia(stack);
     }
 
     @Override
-    public void setMedia(int media) {
+    public void setMedia(long media) {
         holder.setMedia(stack, media);
     }
 
@@ -46,12 +46,12 @@ public record CapItemMediaHolder(MediaHolderItem holder,
     }
 
     @Override
-    public int withdrawMedia(int cost, boolean simulate) {
+    public long withdrawMedia(long cost, boolean simulate) {
         return holder.withdrawMedia(stack, cost, simulate);
     }
 
     @Override
-    public int insertMedia(int amount, boolean simulate) {
+    public long insertMedia(long amount, boolean simulate) {
         return holder.insertMedia(stack, amount, simulate);
     }
 }
